// dashboard/src/utils/hasura-api/generated/schemas/remoteRelationshipItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { RemoteRelationshipDefinition } from './remoteRelationshipDefinition';

export interface RemoteRelationshipItem {
  /** Name of the remote relationship */
  name: string;
  definition: RemoteRelationshipDefinition;
}